update s3 service to delete old images

+++ ecommerce-backend/src/main/java/com/aunraza/ecommercebackend/controllers/ProductController.java
@@ -1,20 +1,23 @@
 package com.aunraza.ecommercebackend.controllers;
 
-import com.aunraza.ecommercebackend.dtos.ProductWithCategoryIdDto;
 import com.aunraza.ecommercebackend.models.Category;
 import com.aunraza.ecommercebackend.models.Product;
 import com.aunraza.ecommercebackend.repositories.CategoryRepository;
 import com.aunraza.ecommercebackend.repositories.ProductRepository;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.aunraza.ecommercebackend.s3.S3Service;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 
 @RestController
 @RequestMapping(path = "api/products")
@@ -22,11 +25,12 @@ public class ProductController {
     private ProductRepository productRepository;
     private CategoryRepository categoryRepository;
 
-    private Logger logger = LoggerFactory.getLogger(getClass());
+    private S3Service s3Service;
 
-    public ProductController(ProductRepository productRepository, CategoryRepository categoryRepository) {
+    public ProductController(ProductRepository productRepository, CategoryRepository categoryRepository, S3Service s3Service) {
         this.productRepository = productRepository;
         this.categoryRepository = categoryRepository;
+        this.s3Service = s3Service;
     }
 
     @GetMapping("")
@@ -48,18 +52,37 @@ public class ProductController {
         return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    @PostMapping("")
-    public ResponseEntity<Product> createProduct(@RequestBody ProductWithCategoryIdDto request, Authentication authentication) {
-        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
-        logger.info(userDetails.toString());
+    @PostMapping()
+    public ResponseEntity<Product> createProduct(@RequestParam("file") MultipartFile file,
+                                                 @RequestParam("name") String name,
+                                                 @RequestParam("description") String description,
+                                                 @RequestParam("price") Double price,
+                                                 @RequestParam("categoryId") Integer categoryId,
+                                                 Authentication authentication) {
+        var username = ((UserDetails) authentication.getPrincipal()).getUsername();
 
-        Optional<Category> category = categoryRepository.findById(request.getCategoryId());
+        Optional<Category> category = categoryRepository.findById(categoryId);
         if (category.isPresent()) {
+            String imageUrl;
+            try {
+
+                if (!file.getContentType().equals(MediaType.IMAGE_JPEG_VALUE)
+                        && !file.getContentType().equals(MediaType.IMAGE_PNG_VALUE)) {
+                    throw new IllegalArgumentException("Only JPEG and PNG images are allowed.");
+                }
+                var key = username + "/" + file.getOriginalFilename() + UUID.randomUUID().toString();
+                String fileExtension = StringUtils.getFilenameExtension(file.getOriginalFilename());
+                var keyWithExtension = key + "." + fileExtension;
+                imageUrl = s3Service.putObject(keyWithExtension, file);
+            } catch (Exception e) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
             var newProduct = new Product();
-            newProduct.setName(request.getName());
-            newProduct.setDescription(request.getDescription());
-            newProduct.setPrice(request.getPrice());
+            newProduct.setName(name);
+            newProduct.setDescription(description);
+            newProduct.setPrice(price);
             newProduct.setCategory(category.get());
+            newProduct.setImageUrl(imageUrl);
             var createdProduct = productRepository.save(newProduct);
             return new ResponseEntity<>(createdProduct, HttpStatus.CREATED);
         }
@@ -69,15 +92,44 @@ public class ProductController {
     @PutMapping("/{productId}")
     public ResponseEntity<Product> modifyProduct(
             @PathVariable Integer productId,
-           @RequestBody ProductWithCategoryIdDto request) {
-        Optional<Category> category = categoryRepository.findById(request.getCategoryId());
+            @RequestParam("file") Optional<MultipartFile> file,
+            @RequestParam("name") String name,
+            @RequestParam("description") String description,
+            @RequestParam("price") Double price,
+            @RequestParam("categoryId") Integer categoryId,
+            Authentication authentication) {
+        var username = ((UserDetails) authentication.getPrincipal()).getUsername();
+        Optional<Category> category = categoryRepository.findById(categoryId);
         Optional<Product> foundProduct = productRepository.findById(productId);
         if (category.isPresent() && foundProduct.isPresent()) {
             var modifiedProduct = foundProduct.get();
-            modifiedProduct.setName(request.getName());
-            modifiedProduct.setDescription(request.getDescription());
-            modifiedProduct.setPrice(request.getPrice());
+            String newImageUrl;
+            try {
+                if (file.isPresent()) {
+                    if (!file.get().getContentType().equals(MediaType.IMAGE_JPEG_VALUE)
+                            && !file.get().getContentType().equals(MediaType.IMAGE_PNG_VALUE)) {
+                        throw new IllegalArgumentException("Only JPEG and PNG images are allowed.");
+                    }
+
+                    var imageUrl = foundProduct.get().getImageUrl();
+                    String[] segments = imageUrl.split("amazonaws.com/");
+                    var oldKey = segments[segments.length -1];
+                    s3Service.deleteObject(oldKey);
+
+                    var newKey = username + "/" + file.get().getOriginalFilename() + UUID.randomUUID().toString();
+                    String fileExtension = StringUtils.getFilenameExtension(file.get().getOriginalFilename());
+                    var keyWithExtension = newKey + "." + fileExtension;
+                    newImageUrl = s3Service.putObject(keyWithExtension, file.get());
+                    modifiedProduct.setImageUrl(newImageUrl);
+                }
+            } catch (Exception e) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
+            modifiedProduct.setName(name);
+            modifiedProduct.setDescription(description);
+            modifiedProduct.setPrice(price);
             modifiedProduct.setCategory(category.get());
+
             productRepository.save(modifiedProduct);
             return new ResponseEntity<>(HttpStatus.OK);
         }
@@ -86,7 +138,19 @@ public class ProductController {
 
     @DeleteMapping("/{productId}")
     public ResponseEntity<Object> deleteProduct(@PathVariable Integer productId) {
-        productRepository.deleteById(productId);
-        return new ResponseEntity<>(HttpStatus.OK);
+        var optionalProduct = productRepository.findById(productId);
+        if (optionalProduct.isPresent()) {
+            var imageUrl = optionalProduct.get().getImageUrl();
+            String[] segments = imageUrl.split("amazonaws.com/");
+            var key = segments[segments.length -1];
+            try {
+                s3Service.deleteObject(key);
+            } catch (Exception e) {
+                return new ResponseEntity<>("S3 fail to delete object", HttpStatus.BAD_REQUEST);
+            }
+            productRepository.deleteById(productId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>("Product with id cannot be found", HttpStatus.NOT_FOUND);
     }
 }
